server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
distlib==0.3.6
fastapi>=0.68.0
filelock==3.9.0
h11==0.14.0
idna==3.10
platformdirs==3.1.1
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
sqlalchemy>=1.4.23
starlette==0.45.3
typing_extensions==4.12.2
uvicorn>=0.15.0
virtualenv==20.20.0
python-dotenv==1.0.0
PyJWT
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator>=2.0.0
cryptography==42.0.2
fastapi-mail==1.4.1
aiosmtplib>=1.0.7
slowapi>=0.1.8
pycryptodome>=3.19.0
